Nostr notifications
